Change rb_rvargc_payload_head to take a void *
I want a function that given a pointer will return the T_PAYLOAD object
associated with that pointer.  Currently we are allocating T_PAYLOAD
after a class, but the "get payload head" function assumed that the
payload object immediately follows.  This happens to be true today, but
it won't be true in the case of `realloc` for example.