Fix iteration in GC
This makes heap iteration a little bit slower, but shouldn't be a big
deal.  Instead of handing out pointers to the start and end of a page,
scan the page looking for T_PAYLOAD objects.  Then hand use the
T_PAYLOAD address as the "end" of the page.  Then walk past any slots
used for the T_PAYLOAD and resume iteration.

`make btest` is green at this point.