Simplify searching for a contiguous space
This commit changes the search algorithm to look ahead when traversing
the linked list.  If the *next* link in the LL is not on the same page,
then we know to advance and check the next segment.

Next we change the page search algorithm to execute GC and find a page
to allocate out of.

NOTE: this has a huge bug in it.  If the heap is perfectly fragmented,
we can get stuck in an infinite loop looking for segments to allocate
out of.  We need to fix this by detecting the problem and allocating a
page.  Maybe we just bump the allocatable pages up by one?  We need to
investigate

Test program looks like this:

```ruby
list = []
loop do
  list << Class.new
  p list.length
end
```