fix T_PAYLOAD acccounting during page sweep
We can't "mark" marking bits in this function because the free loop is
"not"ing and copying the not.  So marking them won't impact objects that
are on the same mark plane.  Instead just leave as-is.  Freeing a
T_PAYLOAD will add the payload and body objects back to the freelist,
and when it does that those objects will be converted in to T_NONE
objects.  The sweep loop already knows how to handle T_NONE objects, so
we can let the sweep loop "naturally" do accounting for those slots.