stop advancing cursor at the end of teh free region
We were always seeing heap pages that contained a T_NONE after every
correctly sized allocation block for a payload. This was because we were
advancing (and returning) the search cursor based on a lookeahead value,
even when we were in the last iteration of the loop (ie. we'd already
found a region).

This extra lookahead block was then never being added back to the
freelist, because we just join together the slots at either end of the
found region (which includes our sad, lonely T_NONE).

We shouldn't advance the cursor in this case to avoid leaving gaps in
the heap page, that are never filled.