Don't double count slots following T_PAYLOAD
Slots following a T_PAYLOAD were being double counted when they were
freed.  The T_PAYLOAD branch of the case statement would count the
length of the T_PAYLOAD as "freed".  But when it iterated the payload
and added them to the freelist, they got set to T_NONE.  The next
iteration of the sweep loop would visit the T_NONE branch of the case
statement and count the slot again.

This commit tags the slots following the T_PAYLOAD with a header bit so
the T_NONE branch knows that the slot has already been counted as part
of the T_PAYLOAD free section.