YJIT: Some debug code to verify possible race with MJIT's worker
MJIT sometimes crashes on CI. While it also crashes on upstream's master
brach, YJIT might be making it worse.

YJIT inserts into the table for mapping direct threading addresses to
opcodes, and that might be racing with mjit_worker trying to compile
things. Without YJIT, that table is constant, so multiple threads trying
to read from it was not a problem.

Use trylock to detect the race. If the race outlined above happens,
trylock should fail to acquire the lock and trigger rb_bug.