T_PAYLOAD objects don't have a klass so don't attempt to mark it.
What's happening here is that we're treating a RPayload like an RBasic
using the RANY macro. so we assume that it has looks like

struct {
  VALUE flags;
  VALUE klass;
}

when it actually looks like

struct {
  VALUE flags;
  unsigned short len;
}

because the data structure members to 64bit boundaries (the `sizeof` the
second struct is 16 bytes, even though it's actually 10) when we ask for
the `klass` field, we actually get back the `len`

This isn't causing any problems for our _current_ implementation,
because we're only using VWA in narrow cases where the len is always
consistent - it's the value 0x03 which, when treated as a VALUE, is the
fixnum 1. Because of this `is_markable_object` returns 0 and the marking
attempt just moves on without incident.

If we change this - either by changing the size of the RPayload struct
so that asking for the second 8 bytes includes some random payload data,
or by changing the size of the payload so the value we're using is no
longer just 0x03 - we cannot guarantee the behaviour of `gc_mark`
because we're effectively going to be taking a random number, treating
it like the address of an RVALUE and attempting to mark whatever's
there.

This commit guards against that