Reduce the length of the RPayload header.
In our current VWA implementation, an RVALUE is followed immediately by
it's payload, which consists of an RPayload struct and then immediately,
it's payload data.

Because the RPayload contains two fields, a VALUE and an unsigned short
- the size gets padded out to be 16 bytes.

This means that when we load an RVALUE (40 bytes) into a 64 bytes cache
line, we'll also pull in the payload header, and then we've only got 8
bytes of useful payload body that's cached (because 64 - 40 - 16 = 8).

This commit shrinks the RPayload struct to 8 bytes, by removing the
second field and instead storing the length in the most significant bits
of the RPayload->flags.

Now we can cache 2 pointers instead of 1.