Give `*` params a name
This commit gives `*` params a name of `*`.  Before this change, a
method like:

```ruby
def foo(*); ->{ super }; end
```

Would not assign a name to the anonymous `*` parameter.  The problem is
that when Ractors try to scan `getlocal` instructions, it puts the name
of the parameter in to a hash.  Since there is no name, we end up with a
strange exception.  This commit gives it a name so that we get the same
exception for `...` as `*`.

Co-Authored-By: John Hawthorn <john@hawthorn.email>