Stop swallowing SEGV messages when the read barrier is present
If the read barrier handler is installed then any segfault will trigger
the read barrier handler. This means that segfaults occuring as as
result of legitimate errors (ie. not as a result of accessing memory
protected with mprotect) will cause a crash with a misleading error
message referencing the read barrier.

GC compaction uses mprotect, so it's possible for any memory that's
managed by the GC to be protected. This commit checks whether the memory
being accessed is part of the GC heap, or not and either calls the read
barrier handler if it is, and falls back to the original handler if not.

This makes debugging memory issues during GC easier, but adds extra
instructions into the segv signal handler. We believe this is acceptable
because:

* The impact this check has on the speed of the sigsegv handler is
massively outweighed by the impact of calling the handler in the first
place, and

* We only use the read barrier intentionally during GC compaction, for
specific cases.

We measured the speed impact of this using the following code:

```
t = Process.clock_gettime(Process::CLOCK_MONOTONIC)
GC.gc_trigger_page_protection_fault
p Process.clock_gettime(Process::CLOCK_MONOTONIC) - t
p GC.stat :read_barrier_faults
```

Where `GC.gc_trigger_page_protection` is defined as follows:

```C
static VALUE
gc_trigger_page_protection_fault(rb_execution_context_t *ec, VALUE self)
{
    rb_objspace_t *objspace = &rb_objspace;

    // install the read barrier handler
    install_handlers();

    for (int i = 0; i < 1000000; i++) {
        // lock the page, remove this line to test unlocked pointer access
        lock_page_body(objspace, GET_PAGE_BODY(self));

        // reading a value from the page should now raise a segv, triggering the handler
        if (!CLASS_OF(self)) { */
            fprintf(stderr, "shouldn't see me\n"); */
        }
    }

    unlock_page_body(objspace, GET_PAGE_BODY(self));
    uninstall_handlers();
    return Qnil;
}
```

What we see is as follows:

* pointer read, unlocked page: `0.00000304`
* pointer read, locked page, no check: `3.3918488579984114`
* Pointer read, locked page, with check: `3.4269131219989504`

As we can see, running the read barrier segv handler for a pointer on a
locked page has slowed down by 1.03%. However, considering that using
mprotect and the read barrier in the first place slows down pointer
access by 112,727,000% on a locked page we believe this performance hit
is acceptable for the sake of easier debugging.