Introduce a size pool inbox
When a string is resized up. We convert it from an embedded string to a
heap allocated string again. At this point it's probably in an
inappropriate size pool.

We should aim to re-embed this strings in a more appropriate size pool
during compaction.

We'll give each size pool an inbox of objects that should be in that
pool. So that when we compact we can drain the inbox to fill slots
before moving the compact cursor.

This commit implements an inbox as a vector of VALUE pointers and when a
string is resized up, it pushed the reference to the string into the
inbox of the most appropriate size pool