[WIP] autoconf opt-in, link with miniruby
Lots are missing and untested. Just from the top of my head:
 - testing out of place build where build dir is not source dir
 - testing BSD make
 - silent cargo prompt
 - mergnig with libruby-static.a and linking with libstatic.{so,bundle}
 - testing what happens if Rust tools are too ol
 - removing the check for libcapstone in configure.ac. Can do it because
   we have optional libcapstone from crates.io.

Adding to MAINLIBS is probably wrong.

This is a start and I've learned enough autoconf and Make for
today.

Note to self: write some notes about the Makefile setup with the various
includes and interactions with the configure script.