Binding generation tool; progress towards compiling simple method (#156)
* Add YJIT bindgen

Mainly for ergonomics. See module level docs for the cruby module for
design considerations.

 * new make target `yjit-bindgen`
 * use autogenerated bindings for two hash related functions

We may choose to depend on the clang crate directly rather than bindgen
as a more programatic approach might be better for this particular tool.

* Export enough so YARV calls YJIT

* Bind enough to get into the per block compile loop

* We can finish generating putnil;leave, but can't do prologue

Made CodeBlock vec a bit bigger so gen_entry_prologue() doesn't bail due
to capacity, but then it trips a todo! when it hits cb.align_pos.
Time to stop and cut a PR. A couple of new bindings; see
bindgen/src/main.rs.

* Some comments

* Use newly learned Relase-Acquire concurrency model

* Add comments and revert `pub` additions

* Fix merge build break and copypasta in C code