Get make yjit-bindgen to work on clean checkout. Fix miniruby build
By adding miniruby as a prerequisite, `make yjit-bindgen` works right
after a fresh `configure` run. This allows simplifying the instruction
for running the command a bit in yjit/README.md.

Tested with `make distclean` and a fresh `configure` run.

Lots of warnings when building miniruby we will fix as we prune more C
YJIT.