Bind enough to be able to trip int3 (#182)
* Bind enough to be able to trip int3

This binds enough code that when you run `./miniruby --yjit` with the
following you run into a SIGTRAP, meaning that miniruby is running
generated code.

    def foo = nil

    i = 0
    while i < 10
      foo
      i += 1
    end

Notice how I'm dancing around stuff that we can't compile yet and how we
curently hardcode the threshold to be 10. Interpreter exits crashes at
the moment.

I had to regenerate the list of opcodes because they were out of sync
with the folllowing

    ./miniruby -e 'puts RubyVM::INSTRUCTION_NAMES.map.with_index{"pub const OP_#{_1.upcase}:usize = #{_2};"}'

I should probably change that part of the build system soonish.

Removing the int3 call reveals a crash that I haven't gotten around to
debugging. Being able to run output code is a good spot to stop for
today.

* minor touchups